[backend] feat(chainig): add variables to scope definition
